requirements.txt
# Core ML/DL Frameworks
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
datasets==2.14.0
tokenizers==0.14.0

# Explanation Methods
captum==0.6.0
shap==0.42.0
lime==0.2.0.1

# Scientific Computing
numpy==1.24.0
pandas==2.1.0
scikit-learn==1.3.0
scipy==1.11.0

# Visualization & UI
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.17.0
streamlit==1.28.0
wordcloud==1.9.2
colorcet==3.0.1

# Computer Vision
opencv-python==4.8.0.76
pillow==10.0.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
isort==5.12.0
flake8==6.0.0
mypy==1.5.0

# Jupyter Notebook Support
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.1.0
notebook==7.0.0

# Utilities
tqdm==4.66.0
pyyaml==6.0.1
click==8.1.0
python-dotenv==1.0.0

# Streamlit Extensions
streamlit-ace==0.1.1
plotly-dash==2.14.0

# Additional ML Libraries
optuna==3.4.0
wandb==0.15.0
tensorboard==2.14.0

# Text Processing
nltk==3.8.1
spacy==3.7.0

# Statistical Analysis
statsmodels==0.14.0
pingouin==0.5.3

# Configuration & Logging
hydra-core==1.3.0
loguru==0.7.0

# HTTP & API
requests==2.31.0
urllib3==2.0.4

# Data Serialization
h5py==3.9.0
joblib==1.3.0
pickle-mixin==1.0.2

# GPU Memory Management (optional)
pynvml==11.5.0

# Code Quality
pre-commit==3.3.0
bandit==1.7.5

# Documentation
sphinx==7.1.0
sphinx-rtd-theme==1.3.0
